+++ tests/integ/modin/series/test_getitem.py
@@ -9,6 +9,7 @@ import pandas as native_pd
 import pytest
 
 import snowflake.snowpark.modin.plugin  # noqa: F401
+from snowflake.snowpark.modin.pandas.utils import try_convert_to_native_index
 from tests.integ.modin.sql_counter import SqlCounter, sql_count_checker
 from tests.integ.modin.utils import eval_snowpark_pandas_result
 
@@ -20,10 +21,10 @@ from tests.integ.modin.utils import eval_snowpark_pandas_result
         [True] * 7,
         [False] * 7,
         np.array([True, True, False, False, False, True, True], dtype=bool),
-        pd.Index([True, True, False, False, False, True, True]),
+        native_pd.Index([True, True, False, False, False, True, True]),
         [True],
         [True, True, False, False, False, True, True, True],
-        pd.Index([], dtype=bool),
+        native_pd.Index([], dtype=bool),
         np.array([], dtype=bool),
     ],
 )
@@ -31,12 +32,15 @@ from tests.integ.modin.utils import eval_snowpark_pandas_result
 def test_series_getitem_with_boolean_list_like(
     key, default_index_snowpark_pandas_series, default_index_native_series
 ):
+    if isinstance(key, native_pd.Index):
+        key = pd.Index(key)
+
     def getitem_helper(ser):
         # Native pandas can only handle boolean list-likes objects of length = num(rows).
         if isinstance(ser, native_pd.Series):
             # If native pandas Series, truncate the series and key.
             _ser = ser[: len(key)]
-            _key = key[: len(_ser)]
+            _key = try_convert_to_native_index(key)[: len(_ser)]
         else:
             _key, _ser = key, ser
         return _ser[_key]
@@ -141,7 +145,7 @@ def test_series_getitem_with_empty_keys(
     eval_snowpark_pandas_result(
         default_index_snowpark_pandas_series,
         default_index_native_series,
-        lambda ser: ser[key],
+        lambda ser: ser[try_convert_to_native_index(key)],
     )